requirements.txt
Flask==3.1.0
mysql-connector-python==9.1.0
pytest